recommendation/dlrm_v3/requirements.txt
torch==2.8.0
fbgemm_gpu==1.3.0
torchrec==1.3.0
gin_config==0.5.0
pandas==2.3.2
tensorboard==2.20.0
